switch about page with new one
