+++ components/Nav/Nav.module.css
@@ -92,3 +92,7 @@
   overflow: hidden;
   background-color: #e5e5e5;
 }
+
+.navButton :global(div) {
+  padding: 0;
+}
