+++ components/Nav/Nav.tsx
@@ -4,7 +4,7 @@ import useTranslation from 'next-translate/useTranslation'
 import classnames from 'classnames'
 import styles from './Nav.module.css'
 import Settings from '../Forms/Settings/SettingsForm'
-import Button from '../Buttons/Button'
+import TextButton from '../Buttons/TextButton/TextButton'
 import { UserContext } from '../../context/UserContext'
 import Modal from '../Modal/Modal'
 
@@ -76,15 +76,15 @@ const Nav = () => {
             </a>
           </li>
           <li className={styles.divider} />
-          <li className={styles.menuItem}>
-            <Button
+          <li className={classnames(styles.menuItem, styles.navButton)}>
+            <TextButton
               onClick={() => {
                 handleOpen()
                 setIsNavOpen(false)
               }}
             >
               <a>{t('common:nav.settings')}</a>
-            </Button>
+            </TextButton>
             <Modal isOpen={isOpen} onDismiss={handleClose}>
               <Modal.Content>
                 <Settings callbackCloseSettings={handleClose} />
